считается основной платеж

+++ AnnuityStandardTests/AnnuityPayment.cs
@@ -6,7 +6,7 @@ public class AnnuityPaymentOutput : IPaymentOutputBase
 {
     public int PaymentNumber { get; set; }
     public DateTime PaymentDate { get; set; }
-    public double PaymentAmount { get; set; }
+    public decimal PaymentAmount { get; set; }
     public decimal BalanceMainDebt { get; set; }
     public decimal PaymentAmountByPercent { get; set; }
 }
